refactor: move theme-generator into nimbus as subpath entrypoint
Consolidate @commercetools/nimbus-theme-generator into
@commercetools/nimbus/theme-generator as a separate entrypoint.
This avoids the overhead of a separate package (versioning, build
config, install) while keeping culori tree-shakeable — consumers
who don't import the entrypoint never bundle it.

- Add ./theme-generator exports entry to nimbus package.json
- Add theme-generator build entry to vite.config.ts
- Move culori dep from nimbus-theme-generator to nimbus
- Copy source + tests to packages/nimbus/src/theme-generator/
- Update all imports across nimbus, docs app, changeset config
- Remove packages/nimbus-theme-generator/ entirely
- All 32 theme-generator tests pass
