+++ .changeset/config.json
@@ -13,8 +13,7 @@
       "@commercetools/nimbus-tokens",
       "@commercetools/nimbus-icons",
       "@commercetools/nimbus-design-token-ts-plugin",
-      "@commercetools/nimbus-mcp",
-      "@commercetools/nimbus-theme-generator"
+      "@commercetools/nimbus-mcp"
     ]
   ],
   "ignore": ["@commercetools/nimbus-i18n", "color-tokens", "blank-app", "docs"],
