+++ apps/docs/package.json
@@ -15,7 +15,6 @@
     "@commercetools/nimbus": "workspace:^",
     "@commercetools/nimbus-docs-build": "workspace:^",
     "@commercetools/nimbus-icons": "workspace:^",
-    "@commercetools/nimbus-theme-generator": "workspace:^",
     "@emotion/react": "catalog:react",
     "@internationalized/date": "catalog:react",
     "@mdx-js/mdx": "^3.1.1",
