+++ apps/docs/src/components/document-renderer/components/live-code-editor/live-code-editor.tsx
@@ -1,6 +1,6 @@
 import * as NimbusUi from "@commercetools/nimbus";
 import * as icons from "@commercetools/nimbus-icons";
-import * as themeGenerator from "@commercetools/nimbus-theme-generator";
+import * as themeGenerator from "@commercetools/nimbus/theme-generator";
 import {
   Box,
   Menu,
