--- packages/nimbus-theme-generator/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@commercetools/nimbus-theme-generator",
-  "version": "3.0.0",
-  "type": "module",
-  "main": "dist/index.cjs",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "sideEffects": false,
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    },
-    "./package.json": "./package.json"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/commercetools/nimbus.git"
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsup",
-    "typecheck": "tsc --noEmit",
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "culori": "^4.0.1"
-  },
-  "peerDependencies": {
-    "@chakra-ui/react": "catalog:react"
-  },
-  "devDependencies": {
-    "@chakra-ui/react": "catalog:react",
-    "@types/culori": "catalog:",
-    "tsup": "catalog:tooling",
-    "typescript": "catalog:tooling",
-    "vitest": "catalog:tooling"
-  }
-}
