--- packages/nimbus-theme-generator/tsup.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig({
-  entry: ["src/index.ts"],
-  format: ["esm", "cjs"],
-  dts: true,
-  clean: true,
-  external: ["@chakra-ui/react", "@commercetools/nimbus"],
-});
