--- packages/nimbus-theme-generator/vitest.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    include: ["src/**/*.spec.ts"],
-    globals: true,
-  },
-});
